backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
openai
paramiko